requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.1
PyJWT==2.9.0
bcrypt==4.2.0
Faker==30.3.0
requests==2.32.3
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
pymongo==4.8.0
dnspython==2.6.1
Werkzeug==3.0.4
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2